% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/crossValidateClassifier.R
\name{crossValidateClassifier}
\alias{crossValidateClassifier}
\alias{crossValidateClassifier.character}
\alias{crossValidateClassifier.default}
\title{Measure SVM Performance with Leave-One-Out Cross-Validation}
\usage{
crossValidateClassifier(df, num_iter, ..., svm_cost = 1, svm_gamma = "def")

\method{crossValidateClassifier}{default}(df, num_iter, ..., svm_cost = 1,
  svm_gamma = "def")

\method{crossValidateClassifier}{character}(df, num_iter, ..., svm_cost = 1,
  svm_gamma = "def")
}
\arguments{
\item{df}{data frame where rows are samples, column 1 is the sample phenotype
(i.e. C, A, or B), and the remaining columns are values for the classifier probesets.}

\item{num_iter}{integer specifying the number of times to sample (w/ replacement)
from the available row numbers. If numIter equals "full", then the function
\code{\link{loadClassifier}} (with a object of class "overall" will be called
instead (default: "full").}

\item{...}{additional arguments include: None}

\item{svm_cost}{numeric value representing the cost parameter when training an svm model
(default: 1)}

\item{svm_gamma}{numermic value (or "def") representingthe gamma parameter when training
an svm model (default: "def")}
}
\value{
numeric

numeric vector
}
\description{
\code{crossValidateClassifier} returns a number representing the accuracy of the
input classifier after performing leave-one-out cross-validation on an svm model

\code{crossValidateClassifier} returns a numeric vector of accuracies determined by full leave-one-out
cross-validation on an svm model.
}
\details{
This function takes a data frame containing classifier probe values for
all training set samples and performs a leave-one-out cross-validation ("full" or
with replacement and a specified number of iterations) on an svm model to determine
the accuracy of the classifier.

This function takes a list of data frames containing classifier probe values
for all training set samples and performs a full leave-one-out cross-validation (w/o
replacement) of the svm model to determine the accuracy of the classifier.
}
\examples{
\dontrun{
crossValidateClassifier(df, 100)
}
data(genericClassifierData)
crossValidateClassifier(train_data, "full", svm_cost=1, svm_gamma="def")
}

