% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multipleTestClassifierSubfunc.R
\name{getLabels}
\alias{getLabels}
\title{Extract Sample Data from a .txt File}
\usage{
getLabels(filename, col = 1, head = TRUE, id = NULL)
}
\arguments{
\item{filename}{tab-separated .txt file sample data.}

\item{col}{integer specifying the column with sample labels or phenotypes (default: 1).}

\item{head}{logical specifying whether the file contains a header (default TRUE).}

\item{id}{integer specifying a secondary column with sample id values (default: NULL).}
}
\value{
character vector
}
\description{
\code{getLabels} returns a character vector with a value for each sample.
}
\details{
This function extracts sample data (labels or phenotypes) to be attached to the
gene expression data for each sample.
}
\examples{
\dontrun{
getLabels("", id=2)
}
}

