% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loadClassifier.R
\name{loadClassifier}
\alias{loadClassifier}
\alias{loadClassifier.default}
\alias{loadClassifier.noCV}
\alias{loadClassifier.yesCV}
\title{Prepare Test Samples for Classification by Trained Classifier}
\usage{
loadClassifier(obj, ..., data, type, svm_cost = 1, svm_gamma = "def")

\method{loadClassifier}{default}(obj, ..., data, type, svm_cost = 1,
  svm_gamma = "def")

\method{loadClassifier}{yesCV}(obj, ..., data, type, svm_cost = 1,
  svm_gamma = "def")

\method{loadClassifier}{noCV}(obj, ..., data, type, svm_cost = 1,
  svm_gamma = "def")
}
\arguments{
\item{obj}{object of class "yesCV" or "noCV" which determines whether to perform
a leave-one-out cross-validation on the training data.}

\item{...}{additional arguments include: "classifier_name_base", "classifier_name"}

\item{data}{data frame of sample data for testing.}

\item{svm_cost}{numeric value specifying the cost for the svm model (default: 1)}

\item{svm_gamma}{numeric value specifying the gamma for the svm model (default: def)}
}
\value{
list
}
\description{
\code{\link{loadClassifier}} returns a list of the same length as the loaded classifier. If no cross-
validation is performed, each element is a data frame of the individual classifier's results
and scores on each test sample. If a leave-one-out cross-validation is performed, each list
member is a list with a data.frame (same as above) and a numeric vector representing the order
in which the samples were removed during cross-validation.

This function takes a data frame containing classifier values for
all training set samples and performs a leave-one-out cross-validation (w/
replacement and a specified number of iterations) on an svm model to determine
the accuracy of the classifier.

This function takes a data frame containing probe values for
all samples to test and predicts the class using the trained svm classifier.
In addition the score, on a scale from 0 to 100, is also reported.
}
\details{
This function is called to prepare properly formatted test data for classification. After loading the
required dataset (txClassifierPAX), \code{\link{testClassifier}}
are called. This function is called by \code{\link{testClassification}} to perform internal
cross-validation on the classifier.
}
\examples{
\dontrun{
loadClassifier(structure(list(), class="txClassifier"))
}
}

