% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotRangeQC.R
\name{plotRangeQC}
\alias{plotRangeQC}
\title{Plot a Range of Classifier Accuracies}
\usage{
plotRangeQC(df, plot_type, comp_type, samples, genes, ybounds = c(0.6, 1),
  save_dir = NULL)
}
\arguments{
\item{df}{data.frame specifying plot values}

\item{plot_type}{character which can be either "pts" or "box" to specify
the type of plot. If more than one y value exists for a given x value,
then "box" shall be value.}

\item{comp_type}{character vector of 2 elements. This value is used to
create the main title of the plot.}

\item{samples}{integer specifying the number of samples used to generate
accuracy values.}

\item{genes}{integer specifying the number of genes (probesets) used to
search for classifiers.}

\item{ybounds}{optional numeric vector to specify the plot range on the
y-axis (default: c(0.6, 1.0).}

\item{save_dir}{optional character specifying a location for saving the
generated plots (default: NULL).}
}
\description{
Produces either a boxplot or a scatter plot of classifier accuracies
determined by leave one out cross validation of a trained SVM model.
}
\examples{
kRange <- seq(51,100)
accur <- rnorm(50, 0.8, 0.1)
ptsData <- data.frame(fss=kRange,acc=accur)

# Scatter Plot
plotRangeQC(ptsData, "pts", comp_type=3, samples=200, genes=15000)

boxAccur <- lapply(accur, function(x) rnorm(25, x, 0.05))
plotdf <- do.call(rbind, lapply(1:length(kRange), function(x) {
data.frame(fss=rep(kRange[x],length(boxAccur[[x]])),
          acc=boxAccur[[x]])}))
# Boxplot
plotRangeQC(plotdf, "box", comp_type=3, samples=200, genes=15000)
}

