% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multipleTestClassifierSubfunc.R
\name{reportClassifierCalls}
\alias{reportClassifierCalls}
\title{Report the Overall Classifier Results}
\usage{
reportClassifierCalls(calls)
}
\arguments{
\item{calls}{list of 3 elements containing individual classifier results.}
}
\value{
data frame
}
\description{
\code{reportClassifierCalls} returns a data frame with 2 or more columns:
1) classification; 2) score; 3+) additional scores.
}
\details{
This function is called to take individual classifier results and return
an overall result for each sample tested. This function calls
\code{\link{determineFinalClass}}.
}
\examples{
\dontrun{
reportClassifierCalls(data)
}
}

