% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scanClassifierRange.R
\name{scanClassifierRange}
\alias{scanClassifierRange}
\title{Scan a Range of Feature Set Sizes for a Specific Classifier}
\usage{
scanClassifierRange(data_subset, range_pars, comp_type, type = type,
  plotQC = TRUE, ...)
}
\arguments{
\item{data_subset}{data frame of RMA values where samples are columns and
probesets as rows for the subset of training set samples in the specified comparison.}

\item{range_pars}{vector specifying feature set size parameters specific to type of
training specified by 'typeK'.}

\item{comp_type}{character vector of 2 elements that specifies the groups of samples
to keep in the data frame (default: 4).}

\item{plotQC}{logical to generate a plot of the accuracies for the specified range of
feature set sizes (default: TRUE).}

\item{...}{additional arguments include: "svm_cost", "svm_gamma",
"classifier_name_base", "classifier_name".}
}
\value{
list
}
\description{
\code{scanClassifierRange} returns a list of 3 elements: 1) list of classifier
probeset ids; 2) numeric vector of classifier accuracies: 3) list of classifier
probesets training data.
}
\details{
This function is called by \code{\link{createClassifier}} when the option
'typeK' is equal to "range". It is a wrapper for calling
\code{\link{selectFeaturesMRMR}} for each value in the specified range and then
calling \code{\link{plotRangeQC}}.
}
\examples{
\dontrun{
func_output_1 <- scanClassifierRange(data_subset=data, comp_type=c("AR","TX"),
                                     range_pars=c("full", 90, 110, 2))
}
}

