% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multipleTrainClassifierSubfunc.R
\name{selectFeaturesMRMR}
\alias{selectFeaturesMRMR}
\title{Select Features using MRMR}
\usage{
selectFeaturesMRMR(data_subset, feat_count)
}
\arguments{
\item{data_subset}{data frame of RMA values where samples are columns and
probesets as rows for the subset of training set samples in the specified comparison.}

\item{feat_count}{integer that specifies the feature set size to be tested.}
}
\value{
mrmr.ensemble object
}
\description{
\code{selectFeaturesMRMR} returns an object from the MRMR.ensemble function based on
the inputs provided by the function \code{\link{scanClassifierRange}}.
}
\details{
This function is called by \code{\link{createClassifier}} or
\code{\link{scanClassifierRange}}. It is a wrapper for calling
\code{\link[mRMRe]{mRMR.data}} and \code{\link[mRMRe]{mRMR.ensemble}}
}
\examples{
\dontrun{
func_output_1 <- selectFeaturesMRMR(data, feat_count=100)
}
}

