% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multipleTrainClassifierSubfunc.R
\name{subsetComparison}
\alias{subsetComparison}
\title{Subset Training Set Data by Classifier Type}
\usage{
subsetComparison(data_set, comp)
}
\arguments{
\item{data_set}{data frame where rows are samples, column 1 is the
sample phenotype (i.e. C, A, or B), and the remaining columns are discretized
values for the classifier probesets}

\item{comp}{vector of 2 elements that specifies the groups of samples to keep in the
frame.}
}
\value{
data frame
}
\description{
\code{subsetComparison} returns a data frame containing only rows with phenotypes
corresponding to the specified classifier.
}
\details{
This function takes a training set data frame and removes all rows (samples) that
do not have a phenotype associated with the classifier in the specified comparison.
}
\examples{
\dontrun{
datasubset <- subsetComparison(df, c("A", "C"))
}
}

