% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/subsetTestData.R
\name{subsetTestData}
\alias{subsetTestData}
\title{Subset classifier probe RMA values from Test Samples}
\usage{
subsetTestData(d, feat)
}
\arguments{
\item{d}{data frame of normalized RMA values for test samples}

\item{feat}{character vector specifying classifier probesets}
}
\value{
data.frame
}
\description{
\code{subsetTestData} returns a data frame with rows equal to the number of
samples and columns equal to the number of classifier probes
}
\details{
This function loads the list of classifier probes takes only that subset of
RMA values in the test samples.
}
\examples{
\dontrun{
t <- c("PAX")
data_file <- paste0("txClassifier" , t)
subsetTestData(data, colnames(data_file[[1]][,-1]), t)
}
}

