% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testClassification.R
\name{testClassification}
\alias{testClassification}
\title{Run Classification on Test Samples with Known Phenotype}
\usage{
testClassification(data_prefix, type = "PAX", id_data = NULL,
  pheno_col = 1, id_col = 2, output_file_prefix = NULL,
  calc_ts_overall_accuracy = FALSE, ...)
}
\arguments{
\item{data_prefix}{character specificying the name of the file (with or without
".csv") to be loaded for classification. Can also be a preloaded data frame that
is correctly formatted (i.e. phenotypes/labels = column 1, rows = samples,
probe data = columns)}

\item{type}{character of value "PAX" or "BX" specifying the type of samples to be
tested (default: PAX).}

\item{id_data}{character specifying a file that contains additional identifying
information about the test samples. Can also be a preloaded vector(w/names) that
be concatenated to the data (default: NULL).}

\item{pheno_col}{numeric specifying the column in 'id_file' with known sample
phenotypes (default: 1).}

\item{id_col}{numeric specifying the column in 'id_file' with sample ids (default: 2).}

\item{output_file_prefix}{character specifying the beginning of the output filename
which will be added to the time and date stamp and ".csv" (default: NULL).}

\item{calc_ts_overall_accuracy}{logical specifying whether to run leave-one-out cross-
validation on the samples being tested (default: FALSE).}

\item{...}{additional arguments include: "col", "head", "id", "data_start_col",
"probeID_col", "classifier_name", "classifier_name_base", "svm_cost", "svm_gamma"}
}
\value{
list
}
\description{
#'\code{testClassification} returns a list of vectors of 4 elements: 1) sample IDs;
2) classification; 3) score; 4) actual classification.
}
\details{
This function is a wrapper for calling \code{\link{loadClassifier}} and outputs
an extra column to calculate accuracy.
}
\examples{
\dontrun{
testClassification(data_prefix="train_data", calc_ts_overall_accuracy=TRUE)
}
}

