% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testClassifier.R
\name{testClassifier}
\alias{testClassifier}
\alias{testClassifier.data.frame}
\alias{testClassifier.default}
\alias{testClassifier.txClassifier}
\title{Test Classifier on Sample Data}
\usage{
testClassifier(obj, train_data_name, data, type, sub_name = "", svm_cost,
  svm_gamma, ..., train_data, a = 0, b = 100)

\method{testClassifier}{default}(obj, train_data_name, data, type,
  sub_name = "", svm_cost, svm_gamma, ..., train_data, a = 0, b = 100)

\method{testClassifier}{txClassifier}(obj, train_data_name, data, type,
  sub_name = "", svm_cost, svm_gamma, ..., train_data, a = 0, b = 100)

\method{testClassifier}{data.frame}(obj, train_data_name, data, type,
  sub_name = "", svm_cost, svm_gamma, ..., train_data, a = 0, b = 100)
}
\arguments{
\item{obj}{data.frame or txClassifier containing the original training set data.}

\item{train_data_name}{character value representing the name of the classifier.}

\item{data}{data data frame of discretized test sample values.}

\item{sub_name}{character value representing the training set sample to remove if performing
a leave-one-out cross-validation on the training set (default: empty)}

\item{svm_cost}{numeric value specifying the cost for the svm model (default: 1)}

\item{svm_gamma}{numeric value specifying the gamma for the svm model (default: def)}

\item{...}{additional arguments include: "classifier_name_base", "classifier_name"}

\item{train_data}{data frame of discretized training sample values.}

\item{a}{numeric representing the lower bound for the score transformation (default: 0).}

\item{b}{numeric representing the upper bound for the score transformation (default: 100).}
}
\value{
data.frame
}
\description{
\code{testClassifier} returns a data frame with 2 columns: 1) diagnosis; 2) TruGRAF score.
}
\details{
This function is called by \code{\link{loadClassifier}} to run the classifiers on sample
data and return a result (call and score) for each.
}
\examples{
\dontrun{
testClassifier()
}
}

