% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testForExistingClassifier.R
\name{testForExistingClassifier}
\alias{testForExistingClassifier}
\title{Test for Existing Classifier}
\usage{
testForExistingClassifier(type = "PAX",
  classifier_name_base = "txClassifier", classifier_name = NULL)
}
\arguments{
\item{classifier_name_base}{character specifying beginning of the name of the data file
containing the feature set for each classifier. This value will be appended with the type
to make the complete filename of the classifier data object to be loaded (default: txClassifier).}

\item{classifier_name}{character specifying the name of the data file containing the feature
set for each classifier. If this value is NULL, then the classifier_name_base and type will
be used to create a filename on which to search and load the classifier object (default: NULL).}
}
\value{
3-length vector of class types
}
\description{
\code{testForExistingClassifier} returns a vector of boolean values
}
\details{
This function checks for the existence of a loaded list of classifier objects. If no object is
loaded, then the 'txClassifierPAX' dataset from the diagnoseTx package
is loaded. Next, the presence of classifier probesets in each of the slots
in the list is tested. To test new samples, all classifiers must be present.
}
\examples{
type <- c("PAX")
testForExistingClassifier(type)
}

