% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/txClassifierObject.R
\name{txClassifier}
\alias{txClassifier}
\title{txClassifier Constructor Function}
\usage{
txClassifier(classes, data_set, svm_cost = 1, svm_gamma = "def",
  type = "PAX")
}
\arguments{
\item{classes}{character vector of 2 elements that specifies the groups of samples to
keep in the data frame}

\item{data_set}{data frame of RMA values for training set with samples as columns and
probesets as rows (default: NULL).}

\item{svm_cost}{numeric value representing the cost parameter when training an svm model
(default: 1)}

\item{svm_gamma}{numermic value (or "def") representingthe gamma parameter when training
an svm model (default: "def")}

\item{type}{character of value "PAX" or "BX" specifying the type of samples to be
tested (default: PAX).}
}
\value{
object of class txClassifier
}
\description{
\code{\link{txClassifier}} returns an object of class txClassifier.
}
\details{
This function is a constructor for building a txClassifier object. The txClassifier
object is a list of parameters for making and using a trained svm model based on
gene expression data from microarrays.
}
\examples{
\dontrun{
txClassifier <- txClassifier(classes=c("notTX", "TX"), data_set="", sample_labels="",
                             svm_cost=1, svm_gamma=0.05, type="PAX")
}
}

